refactor(state.ts): reorganize imports and types in state.ts for better readability and maintainability
The imports and types in state.ts have been reorganized to improve readability and maintainability of the codebase. This change helps to group related types together and make it easier to understand the structure of the file.
